--- src/main/java/ru/javarush/quest/entity/QuestRepository.java
@@ -1,5 +1,0 @@
-package ru.javarush.quest.entity;
-
-public class QuestRepository {
-
-}
