--- src/main/java/ru/javarush/quest/entity/Role.java
@@ -1,6 +1,0 @@
-package ru.javarush.quest.entity;
-
-public enum Role {
-    USER,
-    ADMIN
-}
