+++ src/main/java/ru/javarush/quest/entity/UserRepository.java
@@ -4,12 +4,8 @@ import java.util.Map;
 import java.util.Optional;
 
 public class UserRepository {
-//    private final UserRepository USER_REPOSITORY = new UserRepository();
     private final Map<Long, User> idToUser;
 
-//    public static UserRepository getInstance() {
-//        return USER_REPOSITORY;
-//    }
     public UserRepository(Map<Long, User> idToUser) {
         this.idToUser = idToUser;
     }
