+++ src/main/java/ru/javarush/quest/servlets/LogoutServlet.java
@@ -7,11 +7,11 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet(name = "restart",value = "/restart")
-public class Restart extends HttpServlet {
+@WebServlet(name = "Logout", value = "/logout")
+public class LogoutServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         req.getSession().invalidate();
-        resp.sendRedirect("/Login");
+        req.getRequestDispatcher("index.jsp").forward(req, resp);
     }
 }
